+++ src/geoworkbench/importers/gs2/__init__.py
@@ -1,0 +1,23 @@
+"""GeoScape II ``.gs2`` container support."""
+
+from .container import (
+    Gs2ContainerError,
+    Gs2ContainerLimits,
+    Gs2ContainerManifest,
+    Gs2Member,
+    Gs2TableSummary,
+    extract_gs2,
+    extract_gs2_table,
+    inspect_gs2,
+)
+
+__all__ = [
+    "Gs2ContainerError",
+    "Gs2ContainerLimits",
+    "Gs2ContainerManifest",
+    "Gs2Member",
+    "Gs2TableSummary",
+    "extract_gs2",
+    "extract_gs2_table",
+    "inspect_gs2",
+]